Move some files for managability.
