+++ src/js/components/perfPages/perfTableOptionsUI.js
