+++ src/js/index.js
@@ -26,8 +26,8 @@ import ErrorBoundary from 'components/error';
 import Footer from 'components/footer';
 import FourOhFour from 'components/404';
 import Header from 'components/header';
-import PerfTable from 'components/perfTable';
-import TestMethodsResultsTable from 'components/testMethodResultsTable';
+import PerfTable from 'components/perfPages/perfTable';
+import TestMethodsResultsTable from 'components/perfPages/testMethodResultsTable';
 import getApiFetch from 'utils/api';
 import reducer from 'store';
 import { logError } from 'utils/logging';
